requirements.txt
alembic==1.7.4
apparmor==3.0.7
appdirs==1.4.4
application-utility==1.3.2
APScheduler==3.6.3
asgiref==3.5.2
astroid==2.11.7
asttokens==2.0.7
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
Beaker==1.11.0
beautifulsoup4==4.11.1
blinker==1.4
Brotli==1.0.9
btrfsutil==5.18.1
CacheControl==0.12.6
cachetools==4.2.2
caffeine-ng==4.0.2
certifi==2020.12.5
cffi==1.15.1
chardet==5.0.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
configobj==5.1.0.dev0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.5
distro==1.7.0
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.7.0
django-heroku==0.3.1
django-phonenumber-field==6.4.0
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
drf-yasg==1.20.0
duplicity==0.8.23
email-validator==1.2.1
ewmh==0.1.6
executing==0.10.0
extras==1.0.0
fasteners==0.17.3
filelock==3.6.0
fixtures==4.0.1
Flask==2.1.3
Flask-BabelEx==0.9.4
Flask-Compress==1.12
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
git-bash==0.2
google-api-core==2.8.2
google-api-python-client==2.56.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
greenlet==1.1.2
gssapi==1.7.3
gufw==22.4.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==2.10
importlib-metadata==4.8.1
inflection==0.5.1
ipython==8.4.0
isort==5.10.1
itsdangerous==2.1.2
itypes==1.2.0
jaraco.context==4.1.1
jaraco.functools==3.5.1
jaraco.text==3.8.1
jedi==0.18.1
Jinja2==2.11.3
layoutswitcherlib==0.8.27
lazy-object-proxy==1.7.1
ldap3==2.9
LibAppArmor==3.0.7
lit==14.0.6.dev0
louis==3.22.0
Mako==1.1.6
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==1.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
monotonic==1.6
more-itertools==8.13.0
msgpack==1.0.3
netsnmp-python==1.0a1
npyscreen==4.10.5
nspektr==0.4.0
oauth2client==4.1.3
oauthlib==3.2.0
openapi-codec==1.3.2
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
paramiko==2.11.0
parso==0.8.3
passlib==1.7.4
pbr==5.10.0
pep517==0.12.0
pexpect==4.8.0
phonenumbers==8.12.54
pickleshare==0.7.5
Pillow==8.1.0
pipenv==2022.7.24
platformdirs==2.5.2
ply==3.11
progress==1.6
prompt-toolkit==3.0.30
protobuf==4.21.4
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pulsectl==22.3.2
pure-eval==0.2.2
pwquality==1.4.4
pyalsa==1.2.7
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.21.0
pycodestyle==2.9.0
pycparser==2.21
PyDrive==1.3.1
PyDrive2==1.14.0
Pygments==2.12.0
PyGObject==3.42.2
PyJWT==2.0.1
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
PyNaCl==1.4.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyQt5==5.15.7
PyQt5-sip==12.11.0
PySide6==6.3.1
pysmbc==1.0.23
python-dateutil==2.8.2
python-editor==1.0.4
python-telegram-bot==13.13
python-xlib==0.31
python3-openid==3.2.0
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.28
PyYAML==6.0
ranger-fm==1.9.3
reportlab==3.6.11
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.1
resolvelib==0.8.1
retrying==1.3.3
rich @ file:///build/python-rich/src/rich-12.5.1/dist/rich-12.5.0-py3-none-any.whl
rsa==4.9
ruamel.yaml==0.16.12
setproctitle==1.2.3
shiboken6==6.3.1
simplejson==3.17.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.1
sshtunnel==0.4.0
stack-data==0.4.0
systemd-python==234
team==1.0
tenacity==8.0.1
testtools==2.5.0
tldextract==3.3.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.2
traitlets==5.3.0
trove-classifiers==2022.8.7
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ueberzug==18.1.9
ufw==0.36.1
uritemplate==3.0.1
urllib3==1.26.12
validate==5.1.0.dev0
validate-pyproject==0.9
virtualenv==20.16.2
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.1
